requirements.txt
robotframework==5.0.1
requests==2.27.1